Set Imagenet constants as type float32

+++ happyid/data/config.py
@@ -7,7 +7,7 @@ DIR_BASE = '/kaggle/input/happy-whale-and-dolphin'
 
 NUM_INDIVIDUALS = 15587
 
-MEAN_IMG = np.array([0.485, 0.456, 0.406])
-STD_IMG = np.array([0.229, 0.224, 0.225])
+MEAN_IMG = np.array([0.485, 0.456, 0.406], dtype='float32')
+STD_IMG = np.array([0.229, 0.224, 0.225], dtype='float32')
